add script to store keys
store-key.py is a simple python script that takes a KID and a key
(both specified as hex strings) and creates a key file on the local
filesystem.

This allows playback of a stream that uses that KID, as cencdec will
look for this file, given the KID in the stream.