parse PSSH event from MPD
The PSSH data inside a PSSH event has a different format depending upon
whether it comes from an MPD file or from an MP4 file. In the case of an
MPD file, it is an xml string. In the case of an MP4 file, it is the
binary contents of a pssh box.

This commit adds support for parsing the xml data that is inside a
PSSH event that has been generated by dashdemux.